% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\docType{data}
\name{categories_and_fields}
\alias{categories_and_fields}
\title{Categories and fields}
\format{
A data.frame
}
\usage{
categories_and_fields
}
\description{
This is categories for the fields in the comparison table
}
\keyword{datasets}
