% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\docType{data}
\name{college_chosen_comparison_table}
\alias{college_chosen_comparison_table}
\title{College chosen comparison table}
\format{
A data.frame
}
\usage{
college_chosen_comparison_table
}
\description{
This uses information from NCES to indicate for each college which ones it points to as comparisons, which ones point to it for comparisons (might not be the same), and which ones NCES considers the comparison pool
}
\keyword{datasets}
