% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\docType{data}
\name{comparison_table}
\alias{comparison_table}
\title{Dataset: comparison table
This uses many tables from IPEDS to provide information about colleges over the last ten years.}
\format{
A data.frame. Data for each college for one year is one row.
}
\usage{
comparison_table
}
\description{
Dataset: comparison table
This uses many tables from IPEDS to provide information about colleges over the last ten years.
}
\keyword{datasets}
