% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\docType{data}
\name{population_by_state_by_age}
\alias{population_by_state_by_age}
\title{Populaton by state by age}
\format{
It is a tibble with columns for State_FIPS, State, Age, and Population
}
\usage{
population_by_state_by_age
}
\description{
This uses US Census data for state populations, currently from 2022: https://www2.census.gov/programs-surveys/popest/technical-documentation/file-layouts/2020-2022/sc-est2022-alldata6.pdf
}
\keyword{datasets}
