% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\docType{data}
\name{scorecard_field_organized}
\alias{scorecard_field_organized}
\title{Scorecard field organized}
\format{
A data.frame
}
\usage{
scorecard_field_organized
}
\description{
This is information from college scorecard on earnings by institution by field by degree (and sometimes by gender)
}
\keyword{datasets}
