% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_compare_field_salaries}
\alias{sg_compare_field_salaries}
\title{Return income for graduates from this and similar institutions}
\usage{
sg_compare_field_salaries(institution_id)
}
\arguments{
\item{institution_id}{The UNITID for the institution}
}
\value{
A data.frame of incomes, fields, and colleges from the College Scorecard
}
\description{
Return income for graduates from this and similar institutions
}
