% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_find_college}
\alias{sg_find_college}
\title{Find the college id}
\usage{
sg_find_college(college_name, verbose = TRUE)
}
\arguments{
\item{college_name}{The full college name to match.}

\item{verbose}{Whether to print out similar college names (TRUE or FALSE)}
}
\value{
The ID number for the closest string match
}
\description{
This finds the UNITID number for the relevant college. Enter the full name: "Harvard University" rather than "Harvard". If you choose the verbose option, it will print out the top ten matches (but return only the top one).
}
\examples{
print(sg_find_college("University of Massachusetts, Amherst"))
}
