% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_find_comparisons}
\alias{sg_find_comparisons}
\title{Find comparable colleges}
\usage{
sg_find_comparisons(institution_id)
}
\arguments{
\item{institution_id}{The UNITID for the institution}
}
\value{
A vector of UNITIDs for comparable colleges
}
\description{
Finds colleges that the focal college chooses to compare itself with in federal data as well as ones that the federal government uses as comparisons
}
