% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_open_collegetables}
\alias{sg_open_collegetables}
\title{Open CollegeTables}
\usage{
sg_open_collegetables(institution_id)
}
\arguments{
\item{institution_id}{The UNITID for the institution}
}
\description{
This opens the CollegeTables website for the given institution ID. CollegeTables is a website I made to aggregate info for colleges from various sources and present it in a user-friendly way without paywalls. It was featured in the New York Times here: <https://www.nytimes.com/2023/04/15/your-money/college-cost-data-tools.html>.
}
\details{
This will open the CollegeTables website in your default web browser for a given institution.
}
