% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_return_college_summary}
\alias{sg_return_college_summary}
\title{College summary data}
\usage{
sg_return_college_summary(
  institution_id,
  categories = c("Basic"),
  focal_only = FALSE
)
}
\arguments{
\item{institution_id}{The UNITID for the institution}

\item{categories}{The vector of categories to include. See table(categories_and_fields$Category) for options with number of fields within. Default is "Basic".}

\item{focal_only}{If TRUE, only returns data for the focal institution. Default is FALSE.}
}
\value{
A data.frame with summary data for the focal institution and its comparisons
}
\description{
This function returns a smaller set of summary data from the comparison_table for the focal institution and its comparisons.
}
\details{
The comparison_table data structure has many, many fields (1038 columns in total). This function pulls out a smaller set of data useful for making arguments about shared governance: budgets, enrollments, faculty counts, and so on. By default it will do it for the focal college and its comparisons.
}
\examples{
id <- sg_find_college("University of Massachusetts, Amherst", verbose=FALSE)
summary_data <- sg_return_college_summary(id, categories=c("Basic", "Institution_Finances", "Enrollment"), focal_only=TRUE)
print(summary_data[, c("Year", "Undergrad full time", "Revenue minus expenses")])

}
