% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_return_graduates}
\alias{sg_return_graduates}
\title{Gets the number of graduates over time for the focal institution and comparisons}
\usage{
sg_return_graduates(institution_id, focal_only = FALSE)
}
\arguments{
\item{institution_id}{The UNITID for the institution}

\item{focal_only}{If TRUE, only returns data for the focal institution. Default is FALSE.}
}
\value{
A data.frame of graduates per degree and field per college per year.
}
\description{
This provides aggregated information from IPEDS. It includes information for people with that as a first major and those as a second major (the third column)
}
