\name{sharedgovernance-package}
\alias{sharedgovernance-package}
\alias{sharedgovernance}
\docType{package}
\title{
\packageTitle{sharedgovernance}
}
\description{
\packageDescription{sharedgovernance}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{sharedgovernance}

Maintainer: \packageMaintainer{sharedgovernance}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = sharedgovernance)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{sharedgovernance}}
%\section{Documentation Index}{\packageIndices{sharedgovernance}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
