% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\docType{data}
\name{completions_program}
\alias{completions_program}
\title{Dataset: completions
This uses information originally from tables C2024_A, C2023_A, etc. from IPEDS.
It was processed by code in collegetables.info to organize it and convert the CIPS codes to names of fields.}
\format{
A data.frame with many rows and a few columns. UNITID is the institution ID used to link data.
}
\usage{
completions_program
}
\description{
Dataset: completions
This uses information originally from tables C2024_A, C2023_A, etc. from IPEDS.
It was processed by code in collegetables.info to organize it and convert the CIPS codes to names of fields.
}
\keyword{datasets}
