% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedgovernance.R
\name{sg_compare_field_salaries}
\alias{sg_compare_field_salaries}
\title{Return income for graduates from this and similar institutions}
\usage{
sg_compare_field_salaries(institution_id, focal_only = FALSE)
}
\arguments{
\item{institution_id}{The UNITID for the institution}

\item{focal_only}{If TRUE, only returns data for the focal institution. Default is FALSE.}
}
\value{
A data.frame of incomes, fields, and colleges from the College Scorecard
}
\description{
This provides information from the College Scorecard on earnings by field and degree for the focal institution and its comparisons. For fields with not many graduates, data may be missing.
}
